requirements.txt
amqp==2.4.1
anyjson==0.3.3
astral==1.10.1
attrs==18.2.0
#celery==4.2.1
certifi==2018.11.29
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
#chaussette==1.3.0
CherryPy==18.1.0
#circus==0.15.0
daphne==2.2.5
Django==2.1.7
django-cors-headers==2.4.0
django-filter==2.1.0
django-guardian==1.5.0
django-log-request-id==1.3.2
django-picklefield==2.0
django-rq==1.3.0
#django-supervisor==0.4.0
django-taggit==0.23.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
#flower==0.9.2
gunicorn==19.9.0
jaraco.classes==2.0
Markdown==3.0.1
meld3==1.0.2
msgpack-python==0.5.6
packaging==19.0
Phidget22==1.0.0
PyJWT==1.7.1
pyluach==0.3.0.dev1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
redis==3.1.0
requests==2.21.0
rq==0.13.0
rq-dashboard==0.4.0
rq-scheduler==0.9
rq-scheduler-dashboard==0.0.2
SQLAlchemy==1.2.18
tornado==5.1.1
Twisted==18.9.0
